Add files via upload
So this is kinda the first code i made. It works for me and my friends and I hope it works for you too, might need some tweaking depending on your prefrences.
Just incase you dont care for the readme ill put some instructions here too:
Numpad8 toggles suspend on the code. This can be useful when trying to type in chat because when the hotkey is on, e doesnt exactly work.
Numpad9 to change between slideattacks and autoattacks.
Hold down E for the magic to happen.
Might be required to start in admin.
Feel free to change the keybinds in the code for your comfort.